支持多选题、判断题

+++ modules/exam-service-parent/exam-service-api/src/main/java/com/github/tangyi/exam/api/module/SubjectJudgement.java
@@ -36,7 +36,7 @@ public class SubjectJudgement extends BaseEntity<SubjectJudgement> {
 	 * 分值
 	 */
 	@NotBlank(message = "题目分值不能为空")
-	private Integer score;
+	private Double score;
 
 	/**
 	 * 解析
